pyproject.toml
[project]
name = "openchains"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = "<4.0,>=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "passlib (==1.7.4)",
    "httpx (>=0.28.1,<0.29.0)",
    "tortoise-orm (>=0.24.1,<0.25.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "cryptography (>=44.0.1,<45.0.0)",
    "python-jose[cryptography] (==3.3.0)",
    "bcrypt (==3.2.2)",
    "aerich (>=0.8.2,<0.9.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "tomli-w (>=1.2.0,<2.0.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
]

[tool.poetry]
packages = [
    { include = "backend" },
]

[tool.poetry.scripts]
create-admin = "backend.scripts.create_admin:main_wrapper"
migrate = "backend.scripts.migrate:main_wrapper"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
behave = "^1.2.6"
pytest-asyncio = "^0.25.3"
pre-commit = "^3.7.0"
black = "^24.4.0"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.11.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 88
target-version = [
    "py312",
]
include = "\\.pyi?"

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = ["migration"]

[tool.aerich]
tortoise_orm = "backend.app.db_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
]
build-backend = "poetry.core.masonry.api"
